AI_Project/AI_Resume_Analyzer/requirements.txt
speechRecognition
streamlit
google.generativeai